refactor(web): rename __contracts__ to __core__ folder
Rename features infrastructure folder to avoid confusion with smart
contract terminology in the web3 context.
