+++ apps/web/src/features/walletconnect/index.ts
@@ -7,7 +7,7 @@
  *
  * ```typescript
  * import { WalletConnectFeature } from '@/features/walletconnect'
- * import { useLoadFeature } from '@/features/__contracts__'
+ * import { useLoadFeature } from '@/features/__core__'
  *
  * function MyComponent() {
  *   const walletConnect = useLoadFeature(WalletConnectFeature)
